moteur/L9110_H_bridge/requirements.txt
RPi.GPIO==0.7.0
rpimotorlib==2.7
